+++ renderlib/CMakeLists.txt
@@ -39,8 +39,12 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/Manipulator.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/MathUtil.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/MathUtil.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/MoveTool.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/MoveTool.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Object3d.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Object3d.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Origins.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Origins.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RenderSettings.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RenderSettings.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/RotateTool.cpp"
