+++ renderlib/Manipulator.h
@@ -105,108 +105,6 @@ private:
   std::string m_identifier;
 };
 
-// currently this code is specific to area light but should be tuned to find
-// selected objects in scene and use an average of their centers.
-//
-// TODO create a generic SceneObject class that supports rotation/translation
-// and Origins holds a collection/group of them.  We need to consider whether we
-// are truly rotating the object or in the case of the light source doing some
-// custom version of a rotation (keeping the area light looking toward a fixed target).
-//
-// We still need to support the temporary rotation/translation and then
-// the idea of committing the transform when the user releases the mouse.
-struct Origins
-{
-  enum OriginFlags
-  {
-    kDefault = 0,
-    kNormalize = 1
-  };
-
-  void clear() { m_origins.clear(); }
-  bool empty() const { return m_origins.size() == 0; }
-  void update(SceneView& scene)
-  {
-    if (scene.scene) {
-
-      // e.g. find all selected objects in scene and collect up their centers/transforms here.
-      SceneLight& lt = scene.scene->m_lighting.m_sceneLights[1];
-
-      // save the initial transform? we could use this to reset things if cancelled.
-
-      // this is a copy!!!
-      m_origins = { lt.m_transform };
-      // we want the rotate manipulator to be centered at the target of the light
-      m_origins[0].m_center = lt.m_light->m_Target;
-    }
-  }
-  AffineSpace3f currentReference(SceneView& scene, OriginFlags flags = OriginFlags::kDefault)
-  {
-    return m_origins[0].getAffineSpace();
-  }
-
-  void translate(SceneView& scene, glm::vec3 motion)
-  {
-    glm::vec3 p = m_origins[0].m_center + motion;
-
-    SceneLight& lt = scene.scene->m_lighting.m_sceneLights[1];
-    lt.m_transform.m_center = p;
-
-    lt.Update();
-    scene.renderSettings->m_DirtyFlags.SetFlag(LightsDirty);
-  }
-
-  void rotate(SceneView& scene, glm::quat rotation)
-  {
-    // apply the rotation to the scene's selection
-    // but do not bake it in yet?
-    glm::quat q = rotation * m_origins[0].m_rotation;
-
-    SceneLight& lt = scene.scene->m_lighting.m_sceneLights[1];
-
-    lt.m_transform.m_rotation = q;
-    lt.Update();
-
-    scene.renderSettings->m_DirtyFlags.SetFlag(LightsDirty);
-  }
-
-  std::vector<Transform3d> m_origins;
-};
-
-struct MoveTool : ManipulationTool
-{
-  // Selection codes, are used to identify which manipulator is under the cursor.
-  // The values in this enum are important, lower values means higher picking priority.
-  enum MoveCodes
-  {
-    kMove = 0,
-    kMoveX = 1,
-    kMoveY = 2,
-    kMoveZ = 3,
-    kMoveYZ = 4,
-    kMoveXZ = 5,
-    kMoveXY = 6,
-    kLast = 7
-  };
-
-  MoveTool()
-    : ManipulationTool(kLast)
-    , m_translation(0)
-  {
-  }
-
-  virtual void action(SceneView& scene, Gesture& gesture) final;
-  virtual void draw(SceneView& scene, Gesture& gesture) final;
-
-  // Some data structure to store the initial state of the objects
-  // to move.
-  Origins origins;
-
-  // The current translation of the objects to move.
-  // We need to access this across calls to action and draw
-  glm::vec3 m_translation;
-};
-
 struct AreaLightTool : ManipulationTool
 {
